bot/requirements.txt
aiogram==3.13.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
